+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ To be released
 * Implemented ``/api/study/by/{username}`` under
   ``client.studies.get_by_user`` to get the metadata of all studies
   of a given user.
+* Added ``bots.handle_draw_offer`` and ``bots.handle_takeback_offer`` to handle draw and takeback offers
 
 Thanks to all the contributors who helped to this release:
 - @hsheth2
@@ -25,6 +26,7 @@ Thanks to all the contributors who helped to this release:
 - @MrElyazid
 - @gameroman
 - @JAMoreno-Larios
+- @friedrichtenhagen
 
 
 v0.14.0 (2025-08-26)